Update minimax_ai_player.py
Accounted for the new structure of evaluations where the evaluation function returns two values; the board value, and the neutral incentive value. Restructured part of the code to let three items be passed between layers, instead of two, as the neutral value is now also passed between layers, so that the Negamax algorithm doesn't disincentivise the neutral bonuses intended for both players.